--- tests/unit/test_page_tracker.py
@@ -1,27 +1,0 @@
-''' Unit test for page_tracker'''
-
-import unittest
-
-
-class TestPageTracker(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(cls):
-        pass
-
-    @classmethod
-    def tearDownClass(cls):
-        pass
-
-    def setUp(self):
-        pass
-
-    def tearDown(self):
-        pass
-
-    def test_init_(self):
-        pass
-
-
-if __name__ == "__main__":
-    unittest.main()
